Rename layout -> section
