+++ components/Contact.tsx
@@ -1,10 +1,10 @@
-import Layout from './Layout';
+import Section from './Section';
 import H2 from './H2';
 import ContactForm from './ContactForm';
 import { contact } from '../data';
 
 const Contact = () => ((
-  <Layout bgColor="#d7f2ee" color="#222633">
+  <Section bgColor="#d7f2ee" color="#222633">
     <H2 id="contact" color="#222633" center>{ contact.title }</H2>
     <div className="content">
       <p className="description">{ contact.description }</p>
@@ -23,7 +23,7 @@ const Contact = () => ((
         margin-bottom: 30px;
       }
     `}</style>
-  </Layout>
+  </Section>
 ));
 
 export default Contact;
