+++ components/dapp/SocialSharing.tsx
@@ -1,8 +1,8 @@
-import Layout from '../Layout';
+import Section from '../Section';
 import Content from '../Content';
 
 const SocialSharing = () => (
-  <Layout
+  <Section
     backgroundImage="/static/img-bg-2@2x.jpg"
     bgColor="#1d202b"
     color="#d7f2ee"
@@ -99,7 +99,7 @@ const SocialSharing = () => (
         }
       `}</style>
     </Content>
-  </Layout>
+  </Section>
 );
 
 export default SocialSharing;
